add explicit int conversion and test code for #1
SCFpyr is now tested. And #1 may not be a bug.
In case of the order is odd, reconstructed image is ranged in [0, 255],
however, if the order was even, reconstruction will be over input range.
The normalized reconstructed image, which is covered by [0, 255], seems
to be good and has no difference by odd case one.
Although such behavior was not unexpected, it might have no effect on
except for visualization. #1 will be closed shortly.